スキルテーブルを「スキル名/スキルレベル/経験年数/説明」を表示するように変更、スキルレベルはスマイルアイコン表示で対応、スキルレベル表示のためテーブル定義変更

+++ .doc/ecotiya_com.sql
@@ -83,7 +83,7 @@ CREATE TABLE m_skill
 (
 	skill_field_code varchar(255) NOT NULL COMMENT 'スキル分野コード',
 	skill_name varchar(255) NOT NULL COMMENT 'スキル名称',
-	skill_level decimal(2,1) NOT NULL COMMENT 'スキルレベル',
+	skill_level int(1) NOT NULL COMMENT 'スキルレベル',
 	experience_ym decimal(2,1) NOT NULL COMMENT '経験年数',
 	explanation varchar(255) NOT NULL COMMENT '説明',
 	PRIMARY KEY (skill_field_code, skill_name)
